requirements.txt
imageio==2.9.0
matplotlib==3.4.2
numpy==1.20.3
opencv-python==4.5.3.56
pandas==1.3.2
patchify==0.2.3
pillow==9.0.0
pypng==0.0.21
pyyaml==6.0
pytorch==1.9.0
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.7.1
timm==0.4.12 
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.62.1